requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.3.4
python-multipart==0.0.9
aiosqlite==0.20.0
tenacity==8.5.0
PyJWT==2.8.0
pycryptodome==3.19.0
